+++ app/src/main/java/woowacourse/shopping/common/DefaultAppContainer.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.common
 
-import woowacourse.shopping.data.DefaultCartRepository
-import woowacourse.shopping.data.DefaultProductRepository
+import woowacourse.shopping.data.repository.DefaultCartRepository
+import woowacourse.shopping.data.repository.DefaultProductRepository
 import woowacourse.shopping.repository.CartRepository
 import woowacourse.shopping.repository.ProductRepository
