+++ app/src/main/java/woowacourse/shopping/ui/cart/adapter/CartProductAdapter.kt
@@ -1,8 +1,10 @@
-package woowacourse.shopping.ui.cart
+package woowacourse.shopping.ui.cart.adapter
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.model.Product
+import woowacourse.shopping.ui.cart.DateFormatter
+import woowacourse.shopping.ui.cart.viewholder.CartProductViewHolder
 
 class CartProductAdapter(
     items: List<Product>,
