+++ app/src/main/java/woowacourse/shopping/ui/cart/viewholder/CartProductViewHolder.kt
@@ -1,10 +1,11 @@
-package woowacourse.shopping.ui.cart
+package woowacourse.shopping.ui.cart.viewholder
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemCartProductBinding
 import woowacourse.shopping.model.Product
+import woowacourse.shopping.ui.cart.DateFormatter
 
 class CartProductViewHolder(
     private val binding: ItemCartProductBinding,
