+++ app/src/main/java/woowacourse/shopping/ui/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui
+package woowacourse.shopping.ui.main
 
 import android.content.Intent
 import android.os.Bundle
@@ -10,6 +10,7 @@ import woowacourse.shopping.R
 import woowacourse.shopping.common.CommonViewModelFactory
 import woowacourse.shopping.databinding.ActivityMainBinding
 import woowacourse.shopping.ui.cart.CartActivity
+import woowacourse.shopping.ui.main.adapter.ProductAdapter
 
 class MainActivity : AppCompatActivity() {
