+++ app/src/main/java/woowacourse/shopping/ui/main/adapter/ProductAdapter.kt
@@ -1,12 +1,13 @@
-package woowacourse.shopping.ui
+package woowacourse.shopping.ui.main.adapter
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.model.Product
+import woowacourse.shopping.ui.main.viewholder.ProductViewHolder
 
 class ProductAdapter(
     private val items: List<Product>,
-    private val onClickProduct: (Product) -> Unit
+    private val onClickProduct: (Product) -> Unit,
 ) : RecyclerView.Adapter<ProductViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ProductViewHolder {
